# docs/swagger.yaml
basePath: /api
definitions:
...
  cmd.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  cmd.Voucher:
    properties:
      active:
        type: boolean
      code:
        type: string
      createdAt:
        type: string
      id:
        type: string
      maxUsage:
        type: integer
      type:
        $ref: '#/definitions/go-ddd-clean_internal_domain_voucher.Type'
      unit:
        $ref: '#/definitions/go-ddd-clean_internal_domain_voucher.Unit'
      usedCount:
        type: integer
      validFrom:
        type: string
      validTo:
        type: string
      value:
        format: float64
        type: number
    type: object
  cmd.VoucherCreateRequest:
    properties:
      active:
        type: boolean
      code:
        type: string
      max_usage:
        type: integer
      type:
        type: string
      unit:
        type: string
      valid_from:
        type: integer
      valid_to:
        type: integer
      value:
        type: number
    type: object
  cmd.VoucherRedeemRequest:
    properties:
      code:
        type: string
      discount:
        type: number
      session_id:
        type: string
      tel:
        type: string
    type: object
  cmd.VoucherRedeemResponse:
    properties:
      redemption:
        $ref: '#/definitions/cmd.VoucherRedemption'
      voucher:
        $ref: '#/definitions/cmd.Voucher'
    type: object
  cmd.VoucherRedemption:
    properties:
      createdAt:
        type: string
      discount:
        format: float64
        type: number
      id:
        type: string
      sessionID:
        type: string
      tel:
        type: string
      voucherID:
        type: string
    type: object
  cmd.VoucherUpdateRequest:
    properties:
      active:
        type: boolean
      max_usage:
        type: integer
      unit:
        type: string
      valid_from:
        type: integer
      valid_to:
        type: integer
      value:
        type: number
    type: object
...
  go-ddd-clean_internal_domain_voucher.Type:
    enum:
    - discount
    - free
    type: string
    x-enum-varnames:
    - TypeDiscount
    - TypeFree
  go-ddd-clean_internal_domain_voucher.Unit:
    enum:
    - percent
    - baht
    - session
    type: string
    x-enum-varnames:
    - UnitPercent
    - UnitBaht
    - UnitSession
info:
  contact: {}
  description: API documentation for the go-ddd-clean service.
  title: Photobooth Platforms API
  version: "1.0"
paths:
...
  /api/vouchers:
    get:
      parameters:
      - description: ดึงเฉพาะที่เปิดใช้งาน
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cmd.Voucher'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดึงรายการคูปอง
      tags:
      - Vouchers
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลคูปอง
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.VoucherCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.Voucher'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: สร้างคูปองใหม่
      tags:
      - Vouchers
  /api/vouchers/{id}:
    delete:
      parameters:
      - description: รหัสคูปอง
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ลบคูปอง
      tags:
      - Vouchers
    get:
      parameters:
      - description: รหัสคูปอง
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Voucher'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดูข้อมูลคูปอง
      tags:
      - Vouchers
    put:
      consumes:
      - application/json
      parameters:
      - description: รหัสคูปอง
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลที่ต้องแก้ไข
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.VoucherUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Voucher'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ปรับปรุงคูปอง
      tags:
      - Vouchers
  /api/vouchers/code/{code}:
    get:
      parameters:
      - description: รหัสคูปอง
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Voucher'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ค้นหาคูปองด้วยรหัส
      tags:
      - Vouchers
  /api/vouchers/redeem:
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลการใช้งานคูปอง
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.VoucherRedeemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.VoucherRedeemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: ใช้งานคูปอง
      tags:
      - Vouchers
...
schemes:
- http
securityDefinitions:
  BoothTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"